Interview: implemented queue of disabled questions calculation (fixes test, but supplies incorrect values) KP-2685
Former-commit-id: 93921fdc845f3ae176b7506ac50270f2fe8a2e17

+++ src/Tests/WB.Core.SharedKernels.DataCollection.Tests/InterviewTests/when_answering_question_A_and_that_answer_disables_question_B_and_disabled_B_disables_question_C.cs
@@ -18,7 +18,7 @@ namespace WB.Core.SharedKernels.DataCollection.Tests.InterviewTests
     {
         Establish context = () =>
         {
-            emptyRosterVector = new int[]{};
+            emptyRosterVector = new decimal[]{};
             userId = Guid.Parse("11111111111111111111111111111111");
             Guid questionnaireId = Guid.Parse("77778888000000000000000000000000");
 
@@ -81,6 +81,6 @@ namespace WB.Core.SharedKernels.DataCollection.Tests.InterviewTests
         private static Guid questionCId;
         private static Interview interview;
         private static Guid userId;
-        private static int[] emptyRosterVector;
+        private static decimal[] emptyRosterVector;
     }
 }
